Add logger

+++ src/idea/idea.service.ts
@@ -1,9 +1,9 @@
-import { Injectable } from '@nestjs/common';
-import { Repository } from 'typeorm';
-
-import { IdeaEntity } from './idea.entity';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
 import { IdeaDTO } from './idea.dto';
+import { IdeaEntity } from './idea.entity';
+
 
 @Injectable()
 export class IdeaService {
@@ -22,16 +22,31 @@ export class IdeaService {
   }
 
   async read(id: string) {
-    return await this.ideaRepository.findOne({ id });
+    const idea =  await this.ideaRepository.findOne({ id });
+    if (!idea) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+    return idea;
   }
 
   async update(id: string, data: Partial<IdeaDTO>) {
+
+    let idea = await this.ideaRepository.findOne({where: {id}});
+    if (!idea) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.update({ id }, data);
-    return this.ideaRepository.findOne({ id });
+    idea = await this.ideaRepository.findOne({id});
+    return idea;
   }
 
   async destroy(id: string) {
+
+    const idea = await this.ideaRepository.findOne({where: {id}});
+    if (!idea) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
     await this.ideaRepository.delete({ id });
-    return { deleted: true };
+    return idea;
   }
 }
